\name{JavaGUI-package}
\alias{JavaGUI-package}
\docType{package}
\title{
Guide and useful libraries for building a Java based GUI under R
}
\description{
Guide and useful libraries for building a Java based GUI under R
}
\details{
\tabular{ll}{
Package: \tab JavaGUI\cr
Type: \tab Package\cr
%Version: \tab 0.1\cr
%Date: \tab 2010-08-03\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Since the JavaGUI package has relatively big library files (~ 7MByte)
and should rarely be updated to save space in the CRAN archive,
the more dynamic accompanying guide can be found at \url{http://algorithm-forge.com/JavaGUI} 

}
\author{
Kornelius Rohmeyer

Maintainer: Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\note{
We use the following Java libraries:
\itemize{
 \item Apache Commons Logging under the Apache License, Version 2.0, January 2004, \url{http://commons.apache.org/logging/}, Copyright 2001-2007 The Apache Software Foundation 
 \item Apache jog4j under Apache License 2.0, \url{http://logging.apache.org/log4j/}, Copyright 2007 The Apache Software Foundation
 \item Apache Commons Lang under  Apache License 2.0, \url{http://commons.apache.org/lang/}, Copyright 2001-2011 The Apache Software Foundation
 \item JLaTeXMath under GPL >= 2.0, \url{http://forge.scilab.org/index.php/p/jlatexmath/}, Copyright 2004-2007, 2009 Calixte, Coolsaet, Cleemput, Vermeulen and Universiteit Gent
 \item JRI under Lesser General Public License (LGPL) 2.1, \url{http://www.rforge.net/rJava/}, Copyright 2004-2007 Simon Urbanek 
 \item iText 2.1.4 under LGPL, \url{http://itextpdf.com/}, Copyright by Bruno Lowagie 
 \item SwingWorker under LGPL, \url{https://swingworker.dev.java.net/}, Copyright (c) 2005 Sun Microsystems 
 \item JXLayer under BSD License, \url{https://jxlayer.dev.java.net/}, Copyright 2006-2009, Alexander Potochkin 
 \item JGoodies Forms under BSD License, \url{http://www.jgoodies.com/freeware/forms/}, Copyright JGoodies Karsten Lentzsch  
 \item AFCommons under BSD License, \url{http://www.algorithm-forge.com/afcommons/}, Copyright (c) 2007-2010 by Kornelius Rohmeyer and Bernd Bischl
 \item JHLIR under BSD License, \url{http://jhlir.r-forge.r-project.org/}, Copyright (c) 2008-2010 by Bernd Bischl and Kornelius Rohmeyer  
 \item SGTK under BSD License, \url{http://www.algorithm-forge.com/rjavaclient/}, Copyright (c) 2007-2010 by Bernd Bischl and Kornelius Rohmeyer 
 \item Apache Commons Collections
 \item Apache Commons Validator
 \item MySQL-Connector
 \item Apache POI
 \item SwingX
 \item Wizard
 } 
}
\keyword{ package }
\keyword{ graphs }
\keyword{ htest }

\examples{

\dontrun{
showGuideForJavaGUIs()

}
}
