\name{showGuideForJavaGUIs }
\alias{showGuideForJavaGUIs }

\title{
	Graphical User Interface for the Creation of SRMTP Graphs
}
\description{
	Starts a graphical user interface for the creation of SRMTP graphs.
}
\usage{
	showGuideForJavaGUIs ()
}

\details{
	See the vignette of this package for further details, since describing a GUI interface is better done with a lot of nice pictures.
}
\value{
	The function itself returns NULL.
}
\author{
	Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}

\examples{

\dontrun{
showGuideForJavaGUIs() 

}
}
\keyword{ misc }


